+++ lib/solver_config.cpp
@@ -379,7 +379,10 @@ SolverConfig SolverConfig::load(const string& filename) {
           } else if (ai->id() == "needsPathsFile") {
             sc._needsPathsFile = get_bool(ai);
           } else if (ai->id() == "tags") {
-            sc._tags = get_string_list(ai);
+            std::vector<std::string> loaded = get_string_list(ai);
+          	std::vector<std::string> tags;
+           	std::copy_if (loaded.begin(), loaded.end(), std::back_inserter(tags), [](std::string s){return s != "default";} );
+            sc._tags = tags;
           } else if (ai->id() == "stdFlags") {
             sc._stdFlags = get_string_list(ai);
           } else if (ai->id() == "requiredFlags") {
@@ -575,6 +578,13 @@ void SolverConfigs::addConfig(const MiniZinc::SolverConfig& sc) {
     }
   }
   sc_tags.push_back(name);
+
+  // Add default tag if the solver is marked as the overall default
+  DefaultMap::const_iterator def_it = _tagDefault.find("");
+  if (def_it != _tagDefault.end() && def_it->second == id) {
+		sc_tags.emplace_back("default");
+	}
+
   for (const auto& t : sc_tags) {
     auto it = _tags.find(t);
     if (it == _tags.end()) {